fix: resolve JSCPD and textlint CI failures
- Add src/types/drill.ts with shared DrillData interface to eliminate
  duplicate type definitions between gatsby-node.ts and generateDrillPdf.ts
- Refactor gatsby-node.ts: import DrillData from shared types, extract
  loadDrillsFromDirectory() helper used by both createPages and sourceNodes
- Re-export DrillData from generateDrillPdf.ts for backward compatibility
- Add .github/linters/.jscpd.json: threshold 5%, ignore YAML drill/template
  files (intentionally similar by schema design)
- Fix .textlintrc: restore 'terminology: true' (empty rules object causes
  textlint to exit with error)
