+++ .github/linters/.textlintrc
@@ -2,5 +2,7 @@
   "filters": {
     "comments": true
   },
-  "rules": {}
+  "rules": {
+    "terminology": true
+  }
 }
